refactor: migrate server to the separate file
